+++ src/utils.js
@@ -10,6 +10,11 @@ var utils = {
           newObj[key] = obj[key];
         }
       }
+      if(utils.isObject(obj) && utils.isModel(obj) && ('clientFields' in obj)){
+        obj.clientFields().forEach(function(fld){
+          newObj[fld] = obj[fld];
+        });
+      }
     }
     return newObj;
   },